Fix compiler warnings for C23 compatibility
- Fix typedef union PCode to include type name in cgen.h
- Add function prototypes for rindex() and badFile() in xbitmap.c
- Fix SlotStruct pointer type in objs.c array initialization
- Add function prototypes for HashTable->get and HashTable->remove methods
- Add function prototypes for ClassInfo->slotGetMeth and slotSetMeth
- Fix validObjectP macro to cast pointer to long for HashTable operations
- Fix all HashTable method calls to explicitly cast pointers to long
- Change slotGetMeth/slotSetMeth prototypes from void* to long for correct type matching

+++ src/viola/obj.h
@@ -42,7 +42,7 @@ extern long securityMode;
 #define appendVObjListNode(olist, obj) appendObjToList(olist, obj)
 
 /*int validObjectP();*/
-#define validObjectP(self) objObj2ExistP->get(objObj2ExistP, self)
+#define validObjectP(self) objObj2ExistP->get(objObj2ExistP, (long)(self))
 
 VObj* findObject(long objNameStrID);
 VObjList* appendObjToList();
